requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Audio processing
librosa>=0.10.0
demucs>=4.0.0

# Video processing
moviepy>=1.0.3
imageio>=2.31.0

# Text processing
python-Levenshtein>=0.21.0

# Configuration and data handling
pyyaml>=6.0.0
python-dotenv>=1.0.0

# AI/ML
openai>=1.0.0
anthropic>=0.34.0
google-genai>=0.6.0

# Download utilities
yt-dlp>=2023.11.16
# Spotify API support for fetching track lists and YouTube links
spotipy>=2.25.0
requests>=2.28.0
